+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## v5.4.0 - 2025-03-14
+* Added position label
+* Added summary text for staff & senior engineer positions
+* Added position switcher to change position and summary
+* Refactored html manipulation methods
+* Changed to staff engineer by default
+
 ## v5.3.0 - 2025-03-14
 * Added additional sizes (medium, large)
 * Set medium size to default